test: 현재 참여 인원 API 관련 메서드 이동
